requirements.txt
fastapi==0.115.6
pydantic==2.10.4
openai==1.58.1
google-api-python-client==2.156.0
spacy==3.8.3
uvicorn==0.34.0
PyYAML==6.0.2
pydantic_settings==2.7.0
aiohttp==3.11.11
anthropic==0.42.0
